+++ src/typechecker.c
@@ -15,13 +15,13 @@ char *type_node_text(Node *type) {
   static size_t offset = 0;
   char *outstring = type_node_text_buffer + offset;
   for (unsigned i = 0; i < type->pointer_indirection; ++i) {
-    offset += snprintf(type_node_text_buffer + offset,
-                       TYPE_NODE_TEXT_BUFFER_SIZE - offset,
-                       "@");
+    offset += (size_t) snprintf(type_node_text_buffer + offset,
+      TYPE_NODE_TEXT_BUFFER_SIZE - offset,
+      "@");
   }
-  offset += snprintf(type_node_text_buffer + offset,
-                     TYPE_NODE_TEXT_BUFFER_SIZE - offset,
-                     "%s", type->value.symbol);
+  offset += (size_t) snprintf(type_node_text_buffer + offset,
+    TYPE_NODE_TEXT_BUFFER_SIZE - offset,
+    "%s", type->value.symbol);
   offset += 1;
   if (offset >= TYPE_NODE_TEXT_BUFFER_SIZE) {
     offset = 0;
@@ -497,8 +497,8 @@ Error typecheck_expression
       if (err.type) { return err; }
       err = parse_get_type(context, expression_type_no_pointer, expression_type_info);
       if (err.type) { return err; }
-      cast_type_size = cast_type_info->children->value.integer;
-      expression_type_size = expression_type_info->children->value.integer;
+      cast_type_size = (size_t) cast_type_info->children->value.integer;
+      expression_type_size = (size_t) expression_type_info->children->value.integer;
       free(cast_type_info);
       free(expression_type_info);
